REfactor exposure code to ahve sub-package
- akd/_base/exposure/
