+++ akd/_base/exposure/core.py
@@ -1,9 +1,6 @@
 import sys
 import types
-from dataclasses import dataclass
-from dataclasses import field as dc_field
 from dataclasses import replace as dc_replace
-from enum import StrEnum
 from typing import (
     Any,
     Callable,
@@ -18,15 +15,9 @@ from typing import (
 from loguru import logger
 from pydantic import BaseModel
 
+from .structures import ExposedParam, ExposedParamRuntimeInfo, ExposedParamTypeSource
+
 _EXPOSED_META_VAR_NAME = "_exposed_meta"
-
-
-class ExposedParamTypeSource(StrEnum):
-    FGET = "fget"
-    FSET = "fset"
-    ANNOTATED = "annotated"
-    RUNTIME = "runtime"
-    NONE = "none"
 
 
 def get_type_from_property(prop: property) -> tuple[Any | None, ExposedParamTypeSource]:
@@ -63,154 +54,7 @@ def get_type_from_property(prop: property) -> tuple[Any | None, ExposedParamType
     return None, ExposedParamTypeSource.NONE
 
 
-@dataclass
-class ExposedParam:
-    """Metadata for exposed parameters.
-
-    This model supports both decorator-based (@exposed_param) and annotation-based
-    (Annotated[T, Exposed()]) parameter exposure patterns.
-
-    Attributes:
-        name: Parameter name
-        description: Human readable description
-        extra: Additional metadata for the parameter
-        expose: Whether to create a property for external control
-        persistent: Whether this is persistent state (True) or transient data (False)
-        type_hint: Type hint captured from annotations (e.g., str, int, list[str])
-    """
-
-    name: str = ""
-    description: str = ""
-    expose: bool = True
-    persistent: bool = True
-    type_hint: Any | None = None
-    type_source: ExposedParamTypeSource | None = None
-    extra: dict[str, Any] = dc_field(default_factory=dict)
-
-
-@dataclass
-class ExposedParamRuntimeInfo(ExposedParam):
-    """Complete runtime information about an exposed parameter.
-
-    Inherits name, description, and extra from ExposedParam.
-    Adds runtime-specific fields like type, editability, and current value.
-
-    Attributes:
-        type_: Parameter type as string (e.g., 'str', 'int', 'float')
-        type_source: Source of type inference: 'fget' (getter annotation), 'fset' (setter annotation), 'runtime' (inferred from value), or 'none' (unknown)
-        editable: Whether the parameter has a setter and can be modified
-        current_value: Current runtime value of the parameter (only populated when include_values=True)
-    """
-
-    type_: str = "unknown"
-    editable: bool = False
-    current_value: Any | None = None
-
-
-# Helper functions for Annotated-based exposure
-
-
-def Exposed(
-    name: str | None = None,
-    description: str | None = None,
-    expose: bool = True,
-    **extra,
-) -> ExposedParam:
-    """Create metadata for persistent, exposed config parameters.
-
-    Use this in Annotated type hints to mark config fields that should be:
-    - Exposed as properties on the class
-    - Externally controllable (via UI, API, etc.)
-
-    Args:
-        name: Explicit parameter name (defaults to field name if not provided)
-        description: Human readable description of the parameter
-        **extra: Additional metadata stored in the 'extra' field
-
-    Returns:
-        ExposedParam instance configured for persistent, exposed parameters
-
-    Example:
-        ```python
-        class MyConfig(BaseModel):
-            temperature: Annotated[float, Exposed(
-                description="LLM temperature",
-            )] = 0.7
-        ```
-    """
-    return ExposedParam(
-        name=name or "",
-        description=description or "",
-        expose=expose,
-        persistent=True,
-        extra=extra,
-    )
-
-
-def Validated(
-    name: str | None = None,
-    description: str | None = None,
-    **extra,
-) -> ExposedParam:
-    """Create metadata for transient, validated parameters (input/output schemas).
-
-    Use this for input/output schema fields that should be:
-    - NOT exposed as instance properties (transient data)
-
-    Args:
-        name: Explicit parameter name (defaults to field name if not provided)
-        description: Human readable description of the parameter
-        **extra: Additional metadata stored in the 'extra' field
-
-    Returns:
-        ExposedParam instance configured for transient, validated parameters
-
-    Example:
-        ```python
-        class MyInput(BaseModel):
-            query: Annotated[str, Validated(
-                description="Search query",
-            )]
-        ```
-    """
-    return ExposedParam(
-        name=name or "",
-        description=description or "",
-        expose=False,
-        persistent=False,
-        extra=extra,
-    )
-
-
-def ReadOnly(name: str | None = None, description: str | None = None, expose: bool = True, **extra) -> ExposedParam:
-    """Create metadata for read-only exposed parameters.
-
-    Use this for config fields that should be:
-    - Exposed as properties (readable)
-    - NOT editable (no setter created)
-
-    Args:
-        name: Explicit parameter name (defaults to field name if not provided)
-        description: Human readable description of the parameter
-        **extra: Additional metadata stored in the 'extra' field
-
-    Returns:
-        ExposedParam instance configured for read-only parameters
-
-    Example:
-        ```python
-        class MyConfig(BaseModel):
-            version: Annotated[str, ReadOnly(description="Version info")] = "1.0.0"
-        ```
-    """
-    extra_with_editable = {"editable": False, **extra}
-    return ExposedParam(
-        name=name or "",
-        description=description or "",
-        expose=expose,
-        persistent=True,
-        extra=extra_with_editable,
-    )
+# Property validation and decorator
 
 
 class ValidatedProperty(property):
